backend/wildberries/requirements.txt
accelerate==1.8.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
datasets==3.0.1
DAWG-Python==0.7.2
dill==0.3.8
Django==4.2
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
docopt==0.6.2
drf-yasg==1.21.10
filelock==3.18.0
Flask==3.0.3
frozenlist==1.7.0
fsspec==2024.6.1
huggingface-hub==0.33.1
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
intervaltree==3.1.0
ipymarkup==0.9.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.5.1
multiprocess==0.70.16
natasha==1.6.0
navec==0.10.0
networkx==3.2.1
numpy==1.26.4
packaging==25.0
pandas==2.3.0
propcache==0.3.2
psutil==7.0.0
psycopg2-binary==2.9.10
pyarrow==20.0.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
razdel==0.5.0
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
slovnet==0.6.0
sortedcontainers==2.4.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.52.4
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.5.0
yargy==0.16.0
yarl==1.20.1
zipp==3.23.0
